+++ db/migrate/20260224074819_create_post_comments.rb
@@ -2,8 +2,8 @@ class CreatePostComments < ActiveRecord::Migration[6.1]
   def change
     create_table :post_comments do |t|
       t.text :content
-      t.integer :user_id, null: false
-      t.integer :post_id, null: false
+      t.bigint :user_id, null: false
+      t.bigint :post_id, null: false
       t.datetime :deleted_at
 
       t.timestamps
