+++ db/migrate/20260314022305_create_groups.rb
@@ -3,7 +3,7 @@ class CreateGroups < ActiveRecord::Migration[6.1]
     create_table :groups do |t|
       t.string :name, null: false
       t.text :introduction, null: false
-      t.integer :owner_id, null: false
+      t.bigint :owner_id, null: false
       t.timestamps
     end
   end
