+++ db/migrate/20260314025503_create_group_users.rb
@@ -1,8 +1,8 @@
 class CreateGroupUsers < ActiveRecord::Migration[6.1]
   def change
     create_table :group_users do |t|
-      t.integer :user_id
-      t.integer :group_id
+      t.bigint :user_id
+      t.bigint :group_id
 
       t.timestamps
     end
