+++ db/migrate/20260322035056_create_notifications.rb
@@ -1,11 +1,11 @@
 class CreateNotifications < ActiveRecord::Migration[6.1]
   def change
     create_table :notifications do |t|
-      t.integer :visitor_id, null: false
-      t.integer :visited_id, null: false
-      t.integer :post_id
-      t.integer :comment_id
-      t.integer :group_id
+      t.bigint :visitor_id, null: false
+      t.bigint :visited_id, null: false
+      t.bigint :post_id
+      t.bigint :comment_id
+      t.bigint :group_id
       t.string :action, default: '', null: false
       t.boolean :checked, default: false, null: false
